enhancement(Routing): pluggable routing & StudioApp
