+++ packages/studio-base/src/StudioApp.tsx
@@ -1,0 +1,119 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import { Fragment, Suspense, useEffect } from "react";
+import { DndProvider } from "react-dnd";
+import { HTML5Backend } from "react-dnd-html5-backend";
+import { useOutletContext } from "react-router-dom";
+
+import { AppProps } from "@foxglove/studio-base/App";
+import EventsProvider from "@foxglove/studio-base/providers/EventsProvider";
+import ProblemsContextProvider from "@foxglove/studio-base/providers/ProblemsContextProvider";
+import { StudioLogsSettingsProvider } from "@foxglove/studio-base/providers/StudioLogsSettingsProvider";
+import TimelineInteractionStateProvider from "@foxglove/studio-base/providers/TimelineInteractionStateProvider";
+
+import Workspace from "./Workspace";
+import DocumentTitleAdapter from "./components/DocumentTitleAdapter";
+import MultiProvider from "./components/MultiProvider";
+import PlayerManager from "./components/PlayerManager";
+import SendNotificationToastAdapter from "./components/SendNotificationToastAdapter";
+import StudioToastProvider from "./components/StudioToastProvider";
+import NativeAppMenuContext from "./context/NativeAppMenuContext";
+import NativeWindowContext from "./context/NativeWindowContext";
+import { UserNodeStateProvider } from "./context/UserNodeStateContext";
+import CurrentLayoutProvider from "./providers/CurrentLayoutProvider";
+import ExtensionCatalogProvider from "./providers/ExtensionCatalogProvider";
+import ExtensionMarketplaceProvider from "./providers/ExtensionMarketplaceProvider";
+import PanelCatalogProvider from "./providers/PanelCatalogProvider";
+import { LaunchPreference } from "./screens/LaunchPreference";
+
+// Suppress context menu for the entire app except on inputs & textareas.
+function contextMenuHandler(event: MouseEvent) {
+  if (event.target instanceof HTMLInputElement || event.target instanceof HTMLTextAreaElement) {
+    return;
+  }
+
+  event.preventDefault();
+  return false;
+}
+
+export function StudioApp(): JSX.Element {
+  const props = useOutletContext<AppProps>();
+
+  const {
+    dataSources,
+    extensionLoaders,
+    nativeAppMenu,
+    nativeWindow,
+    deepLinks,
+    enableLaunchPreferenceScreen,
+    extraProviders,
+  } = props;
+
+  const providers = [
+    /* eslint-disable react/jsx-key */
+    <TimelineInteractionStateProvider />,
+    <UserNodeStateProvider />,
+    <CurrentLayoutProvider />,
+    <ExtensionMarketplaceProvider />,
+    <ExtensionCatalogProvider loaders={extensionLoaders} />,
+    <PlayerManager playerSources={dataSources} />,
+    <EventsProvider />,
+    /* eslint-enable react/jsx-key */
+  ];
+
+  if (nativeAppMenu) {
+    providers.push(<NativeAppMenuContext.Provider value={nativeAppMenu} />);
+  }
+
+  if (nativeWindow) {
+    providers.push(<NativeWindowContext.Provider value={nativeWindow} />);
+  }
+
+  if (extraProviders) {
+    providers.unshift(...extraProviders);
+  }
+
+  // The toast and logs provider comes first so they are available to all downstream providers
+  providers.unshift(<StudioToastProvider />);
+  providers.unshift(<StudioLogsSettingsProvider />);
+
+  // Problems provider also must come before other, depdendent contexts.
+  providers.unshift(<ProblemsContextProvider />);
+
+  const MaybeLaunchPreference = enableLaunchPreferenceScreen === true ? LaunchPreference : Fragment;
+
+  useEffect(() => {
+    document.addEventListener("contextmenu", contextMenuHandler);
+    return () => {
+      document.removeEventListener("contextmenu", contextMenuHandler);
+    };
+  }, []);
+
+  return (
+    <MaybeLaunchPreference>
+      <MultiProvider providers={providers}>
+        <DocumentTitleAdapter />
+        <SendNotificationToastAdapter />
+        <DndProvider backend={HTML5Backend}>
+          <Suspense fallback={<></>}>
+            <PanelCatalogProvider>
+              <Workspace
+                deepLinks={deepLinks}
+                appBarLeftInset={props.appBarLeftInset}
+                onAppBarDoubleClick={props.onAppBarDoubleClick}
+                showCustomWindowControls={props.showCustomWindowControls}
+                isMaximized={props.isMaximized}
+                onMinimizeWindow={props.onMinimizeWindow}
+                onMaximizeWindow={props.onMaximizeWindow}
+                onUnmaximizeWindow={props.onUnmaximizeWindow}
+                onCloseWindow={props.onCloseWindow}
+              />
+            </PanelCatalogProvider>
+          </Suspense>
+        </DndProvider>
+      </MultiProvider>
+    </MaybeLaunchPreference>
+  );
+}
