+++ packages/studio-desktop/src/renderer/getDefaultRouter.tsx
@@ -1,9 +1,9 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
-import { createHashRouter } from "react-router-dom";
+import { Navigate, createHashRouter } from "react-router-dom";
 
-import { IAppConfiguration } from "@foxglove/studio-base";
+import { IAppConfiguration, StudioApp } from "@foxglove/studio-base";
 import Root from "@foxglove/studio-desktop/src/renderer/Root";
 
 type Router = ReturnType<typeof createHashRouter>;
@@ -13,9 +13,14 @@ export const getDefaultRouter = async (props: {
 }): Promise<Router> =>
   createHashRouter([
     {
-      path: "*",
+      path: "/",
       element: (
         <Root appConfiguration={props.appConfiguration} extraProviders={[]} dataSources={[]} />
       ),
+      children: [
+        { path: "view", element: <StudioApp /> },
+        { path: "", element: <StudioApp /> },
+      ],
     },
+    { path: "*", element: <Navigate to="/view" /> },
   ]);
